oaknut: Avoid negation of unsigned values

+++ include/oaknut/oaknut.hpp
@@ -110,8 +110,8 @@ public:
             return;
         if (MovImm16::is_valid(imm))
             return MOVZ(wd, imm);
-        if (MovImm16::is_valid(~static_cast<std::uint64_t>(imm)))
-            return MOVN(wd, ~static_cast<std::uint64_t>(imm));
+        if (MovImm16::is_valid(static_cast<std::uint32_t>(~imm)))
+            return MOVN(wd, static_cast<std::uint32_t>(~imm));
         if (detail::encode_bit_imm(imm))
             return ORR(wd, WzrReg{}, imm);
